Add files via upload
This notebook presents the application of the Vector-valued EfficientNet (V-EfficientNet) for the classification of blood smear images from the ALL-IDB2 dataset. Such kind of problem appears in the context of computer-aided diagnosis of acute lymphoblastic leukemia (ALL) and poses some challenges due to the scarcity of labeled images. The paper "V-EfficientNets: Vector-Valued Efficiently Scaled Convolutional Neural Network Models" details the implementation of V-EfficientNet as well as the application of the V-EfficientNetV2-B0 for this classification task.